refactor: move to more standard structure
