+++ example/projects.json
@@ -49,8 +49,8 @@
         "homepage": "https://nextjs.org",
         "default_branch": "canary",
         "created_at": "2016-10-05T23:32:51Z",
-        "pushed_at": "2025-04-25T23:30:01Z",
-        "updated_at": "2025-04-25T23:30:07Z",
+        "pushed_at": "2025-05-03T01:13:58Z",
+        "updated_at": "2025-05-03T00:46:35Z",
         "html_url": "https://github.com/vercel/next.js",
         "clone_url": "https://github.com/vercel/next.js.git",
         "git_url": "git://github.com/vercel/next.js.git",
@@ -58,15 +58,15 @@
         "svn_url": "https://github.com/vercel/next.js",
         "language": "JavaScript",
         "fork": false,
-        "forks_count": 28234,
-        "network_count": 28234,
-        "open_issues_count": 3206,
-        "open_issues": 3206,
-        "stargazers_count": 131365,
-        "subscribers_count": 1479,
-        "watchers_count": 131365,
-        "watchers": 131365,
-        "size": 2229544,
+        "forks_count": 28295,
+        "network_count": 28295,
+        "open_issues_count": 3205,
+        "open_issues": 3205,
+        "stargazers_count": 131537,
+        "subscribers_count": 1482,
+        "watchers_count": 131537,
+        "watchers": 131537,
+        "size": 2236228,
         "organization": {
           "login": "vercel",
           "id": 14985020,
@@ -209,8 +209,8 @@
         "homepage": "https://ai-sdk.dev",
         "default_branch": "main",
         "created_at": "2023-05-23T15:04:08Z",
-        "pushed_at": "2025-04-25T22:34:49Z",
-        "updated_at": "2025-04-25T22:59:00Z",
+        "pushed_at": "2025-05-02T16:27:03Z",
+        "updated_at": "2025-05-03T01:05:51Z",
         "html_url": "https://github.com/vercel/ai",
         "clone_url": "https://github.com/vercel/ai.git",
         "git_url": "git://github.com/vercel/ai.git",
@@ -218,15 +218,15 @@
         "svn_url": "https://github.com/vercel/ai",
         "language": "TypeScript",
         "fork": false,
-        "forks_count": 2136,
-        "network_count": 2136,
-        "open_issues_count": 520,
-        "open_issues": 520,
-        "stargazers_count": 13700,
-        "subscribers_count": 77,
-        "watchers_count": 13700,
-        "watchers": 13700,
-        "size": 29235,
+        "forks_count": 2158,
+        "network_count": 2158,
+        "open_issues_count": 539,
+        "open_issues": 539,
+        "stargazers_count": 13863,
+        "subscribers_count": 78,
+        "watchers_count": 13863,
+        "watchers": 13863,
+        "size": 30642,
         "organization": {
           "login": "vercel",
           "id": 14985020,
@@ -362,8 +362,8 @@
         "homepage": "https://vercel.com/storage",
         "default_branch": "main",
         "created_at": "2022-07-06T15:41:25Z",
-        "pushed_at": "2025-04-24T10:32:49Z",
-        "updated_at": "2025-04-24T08:47:50Z",
+        "pushed_at": "2025-05-02T23:20:01Z",
+        "updated_at": "2025-05-02T13:53:23Z",
         "html_url": "https://github.com/vercel/storage",
         "clone_url": "https://github.com/vercel/storage.git",
         "git_url": "git://github.com/vercel/storage.git",
@@ -371,15 +371,15 @@
         "svn_url": "https://github.com/vercel/storage",
         "language": "TypeScript",
         "fork": false,
-        "forks_count": 68,
-        "network_count": 68,
+        "forks_count": 69,
+        "network_count": 69,
         "open_issues_count": 46,
         "open_issues": 46,
         "stargazers_count": 543,
-        "subscribers_count": 47,
+        "subscribers_count": 48,
         "watchers_count": 543,
         "watchers": 543,
-        "size": 91663,
+        "size": 91672,
         "organization": {
           "login": "vercel",
           "id": 14985020,
@@ -515,7 +515,7 @@
         "default_branch": "main",
         "created_at": "2019-10-28T18:16:01Z",
         "pushed_at": "2025-04-17T10:10:37Z",
-        "updated_at": "2025-04-25T13:13:20Z",
+        "updated_at": "2025-05-03T00:19:23Z",
         "html_url": "https://github.com/vercel/swr",
         "clone_url": "https://github.com/vercel/swr.git",
         "git_url": "git://github.com/vercel/swr.git",
@@ -525,12 +525,12 @@
         "fork": false,
         "forks_count": 1266,
         "network_count": 1266,
-        "open_issues_count": 163,
-        "open_issues": 163,
-        "stargazers_count": 31377,
-        "subscribers_count": 218,
-        "watchers_count": 31377,
-        "watchers": 31377,
+        "open_issues_count": 164,
+        "open_issues": 164,
+        "stargazers_count": 31407,
+        "subscribers_count": 219,
+        "watchers_count": 31407,
+        "watchers": 31407,
         "size": 3504,
         "organization": {
           "login": "vercel",
