+++ example/results/do-they-have-a-readme.csv
@@ -1,5 +1,5 @@
 Project Path,Status,Output
 ./projects/next.js,Success,Yes
-./projects/ai,Success,Yes
+./projects/ai,Success,No
 ./projects/storage,Success,Yes
 ./projects/swr,Success,Yes
