+++ example/results/do-they-have-a-readme.md
@@ -1,6 +1,6 @@
 | Project Path | Status | Output |
 | --- | --- | --- | 
 | ./projects/next.js | Success | Yes |
-| ./projects/ai | Success | Yes |
+| ./projects/ai | Success | No |
 | ./projects/storage | Success | Yes |
 | ./projects/swr | Success | Yes |
