+++ example/results/how-activily-maintained-is-the-project.csv
@@ -1,5 +1,5 @@
 Project Path,Status,Output
 ./projects/next.js,Success,Last commit was 0 days ago.
 ./projects/ai,Success,Last commit was 0 days ago.
-./projects/storage,Success,Last commit was 15 days ago.
-./projects/swr,Success,Last commit was 8 days ago.
+./projects/storage,Success,Last commit was 0 days ago.
+./projects/swr,Success,Last commit was 15 days ago.
