+++ example/results/how-activily-maintained-is-the-project.md
@@ -2,5 +2,5 @@
 | --- | --- | --- | 
 | ./projects/next.js | Success | Last commit was 0 days ago. |
 | ./projects/ai | Success | Last commit was 0 days ago. |
-| ./projects/storage | Success | Last commit was 15 days ago. |
-| ./projects/swr | Success | Last commit was 8 days ago. |
+| ./projects/storage | Success | Last commit was 0 days ago. |
+| ./projects/swr | Success | Last commit was 15 days ago. |
